Some minor fixes.
- Update dependencies
- Fix interlacing
- Increase gradle ram limit
- Fix minor memory leak in in-game menu when closing & reopening.

+++ app/src/main/assets/shaders/frameblend.frag
@@ -8,7 +8,7 @@ out vec4 FragColor;
 
 uniform sampler2D tex;
 uniform sampler2D last_tex;
-uniform float odd_frame;
+uniform int odd_frame;
 uniform bool interlacing;
 uniform bool frameblending;
 
@@ -18,9 +18,9 @@ void main()
 	vec4 new = texture(tex, Texcoord);
 
 	if (interlacing) {
-		float darken = floor(mod(Texcoord.y * 144.0 + odd_frame, 2.0));
+		float darken = floor(mod(Texcoord.y * 144.0 + float(odd_frame), 2.0));
 		new *= darken * 0.5 + 0.5;
-		darken = floor(mod(Texcoord.y * 144.0 + odd_frame + 1.0, 2.0));
+		darken = floor(mod(Texcoord.y * 144.0 + float(odd_frame) + 1.0, 2.0));
 		old *= darken * 0.5 + 0.5;
 	}
 	if (frameblending) {
